pyproject.toml
[project]
name = "notion-ai-mcp"
version = "0.1.0"
description = "MCP Server for Notion API integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.23.1",
    "notion-client>=2.2.1",
    "python-dotenv>=1.0.0",
]
